requirements.txt


Django==5.1.1
djangorestframework==3.14.0
msgpack==1.1.0
mysqlclient==2.2.4
placebo==0.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
